fix: AI v2 gRPC service name, marshallers, unread count — v1.3.0.3
- Fixed messenger.AIService/* → messenger.ChatService/* (all 15 RPC calls)
- Fixed 7 critical marshaller bugs (RateAIAgent, UsageStats, parseAgentInfoV2, etc.)
- Added missing proto fields (AgentInfoV2 fields 15-21, error fields, totalTokensUsed)
- Fixed unread indicator: race condition in ChatListViewModel.loadChats()
- Increased unread highlight alpha for better visibility
- Fixed UsageStat Long → Int type mismatch (server uses int32)

+++ app/src/main/java/lavender/client/android/data/ai/AiV2ChatUseCase.kt
@@ -300,7 +300,7 @@ object AiV2ChatUseCase {
         }
     }
 
-    suspend fun getUsageStats(): Result<Pair<List<UsageStat>, Pair<Long, Int>>> {
+    suspend fun getUsageStats(): Result<Pair<List<UsageStat>, Pair<Int, Int>>> {
         return try {
             val response = RealGrpcClient.aiV2Client.getUsageStats()
             val stats = response.stats.map { it.toDomain() }
